refractor: move VolumeViewer components into their own subfolder
